backend/requirements.txt
Flask
Flask-CORS
redis
pika  # for RabbitMQ
psycopg2-binary # for PostgreSQL